// x-pack/platform/plugins/shared/agent_builder/public/application/components/agents/common/styles.d.ts
export declare const useListDetailPageStyles: () => {
    header: import("@emotion/react").SerializedStyles;
    body: import("@emotion/react").SerializedStyles;
    searchColumn: import("@emotion/react").SerializedStyles;
    searchInputWrapper: import("@emotion/react").SerializedStyles;
    scrollableList: import("@emotion/react").SerializedStyles;
    detailPanelWrapper: import("@emotion/react").SerializedStyles;
    noSelectionPlaceholder: import("@emotion/react").SerializedStyles;
    loadingSpinner: import("@emotion/react").SerializedStyles;
};